+++ documentation/Changelog.md
@@ -167,11 +167,18 @@ Time Copy was only converted into a side-bar version from version 0.8.7.
 - fix: export configs directly after changes
 - other: mini service for session storage
 
-## V 1.0.22 - 1.0.33
+## V 1.0.22 - 1.1.40
 - feature: dlc amag protime - turn ticketnomber pasting into text on and off
 - feature: liquid silver theme
+- feature: profile picture
+- feature: profile file encoding
+- feature: profile options as dropdown
 - fix: small screen styles
 - fix: showAllMessages default value
+- fix: other import/export issues
 - other: restructure and clean up code
 - other: map theme select dynamicaly
 - other: profile cleanup (redesign and select options)
+- other: new profile file version (TCP1)
+- other: render theme and options dynamicaly to html
+- other: profile import type as dynamic attribute to html